+++ pyheos/heos.py
@@ -35,7 +35,6 @@ from pyheos.system import HeosSystem
 from . import command as c
 from . import const
 from .dispatch import Dispatcher
-from .group import HeosGroup
 from .player import HeosNowPlayingMedia, HeosPlayer, PlayMode
 from .types import (
     AddCriteriaType,
@@ -888,194 +887,7 @@ class PlayerMixin(ConnectionMixin):
         return bool(payload[c.ATTR_UPDATE] == c.VALUE_UPDATE_EXIST)
 
 
-class GroupMixin(ConnectionMixin):
-    """A mixin to provide access to the group commands."""
-
-    def __init__(self, *args: Any, **kwargs: Any) -> None:
-        """Init a new instance of the BrowseMixin."""
-        super(GroupMixin, self).__init__(*args, **kwargs)
-        self._groups: dict[int, HeosGroup] = {}
-        self._groups_loaded = False
-
-    @property
-    def groups(self) -> dict[int, HeosGroup]:
-        """Get the loaded groups."""
-        return self._groups
-
-    async def get_groups(self, *, refresh: bool = False) -> dict[int, HeosGroup]:
-        """Get available groups.
-
-        References:
-            4.3.1 Get Groups"""
-        if not self._groups_loaded or refresh:
-            groups = {}
-            result = await self._connection.command(GroupCommands.get_groups())
-            payload = cast(Sequence[dict], result.payload)
-            for data in payload:
-                group = HeosGroup._from_data(data, cast("Heos", self))
-                groups[group.group_id] = group
-            self._groups = groups
-            # Update all statuses
-            await asyncio.gather(
-                *[
-                    group.refresh(refresh_base_info=False)
-                    for group in self._groups.values()
-                ]
-            )
-            self._groups_loaded = True
-        return self._groups
-
-    async def get_group_info(
-        self,
-        group_id: int | None = None,
-        group: HeosGroup | None = None,
-        *,
-        refresh: bool = False,
-    ) -> HeosGroup:
-        """Get information about a group.
-
-        Only one of group_id or group should be provided.
-
-        Args:
-            group_id: The identifier of the group to get information about. Only one of group_id or group should be provided.
-            group: The HeosGroup instance to update with the latest information. Only one of group_id or group should be provided.
-            refresh: Set to True to force a refresh of the group information.
-
-        References:
-            4.3.2 Get Group Info"""
-        if group_id is None and group is None:
-            raise ValueError("Either group_id or group must be provided")
-        if group_id is not None and group is not None:
-            raise ValueError("Only one of group_id or group should be provided")
-
-        # if only group_id provided, try getting from loaded
-        if group is None:
-            assert group_id is not None
-            group = self._groups.get(group_id)
-        else:
-            group_id = group.group_id
-
-        if group is None or refresh:
-            # Get the latest information
-            result = await self._connection.command(
-                GroupCommands.get_group_info(group_id)
-            )
-            payload = cast(dict[str, Any], result.payload)
-            if group is None:
-                group = HeosGroup._from_data(payload, cast("Heos", self))
-            else:
-                group._update_from_data(payload)
-            await group.refresh(refresh_base_info=False)
-        return group
-
-    async def set_group(self, player_ids: Sequence[int]) -> None:
-        """Create, modify, or ungroup players.
-
-        Args:
-            player_ids: The list of player identifiers to group or ungroup. The first player is the group leader.
-
-        References:
-            4.3.3 Set Group"""
-        await self._connection.command(GroupCommands.set_group(player_ids))
-
-    async def create_group(
-        self, leader_player_id: int, member_player_ids: Sequence[int]
-    ) -> None:
-        """Create a HEOS group.
-
-        Args:
-            leader_player_id: The player_id of the lead player in the group.
-            member_player_ids: The player_ids of the group members.
-
-        References:
-            4.3.3 Set Group"""
-        player_ids = [leader_player_id]
-        player_ids.extend(member_player_ids)
-        await self.set_group(player_ids)
-
-    async def remove_group(self, group_id: int) -> None:
-        """Ungroup the specified group.
-
-        Args:
-            group_id: The identifier of the group to ungroup. Must be the lead player.
-
-        References:
-            4.3.3 Set Group
-        """
-        await self.set_group([group_id])
-
-    async def update_group(
-        self, group_id: int, member_player_ids: Sequence[int]
-    ) -> None:
-        """Update the membership of a group.
-
-        Args:
-            group_id: The identifier of the group to update (same as the lead player_id)
-            member_player_ids: The new player_ids of the group members.
-        """
-        await self.create_group(group_id, member_player_ids)
-
-    async def get_group_volume(self, group_id: int) -> int:
-        """
-        Get the volume of a group.
-
-        References:
-            4.3.4 Get Group Volume
-        """
-        result = await self._connection.command(
-            GroupCommands.get_group_volume(group_id)
-        )
-        return result.get_message_value_int(c.ATTR_LEVEL)
-
-    async def set_group_volume(self, group_id: int, level: int) -> None:
-        """Set the volume of the group.
-
-        References:
-            4.3.5 Set Group Volume"""
-        await self._connection.command(GroupCommands.set_group_volume(group_id, level))
-
-    async def group_volume_up(
-        self, group_id: int, step: int = const.DEFAULT_STEP
-    ) -> None:
-        """Increase the volume level.
-
-        References:
-            4.3.6 Group Volume Up"""
-        await self._connection.command(GroupCommands.group_volume_up(group_id, step))
-
-    async def group_volume_down(
-        self, group_id: int, step: int = const.DEFAULT_STEP
-    ) -> None:
-        """Increase the volume level.
-
-        References:
-            4.2.7 Group Volume Down"""
-        await self._connection.command(GroupCommands.group_volume_down(group_id, step))
-
-    async def get_group_mute(self, group_id: int) -> bool:
-        """Get the mute status of the group.
-
-        References:
-            4.3.8 Get Group Mute"""
-        result = await self._connection.command(GroupCommands.get_group_mute(group_id))
-        return result.get_message_value(c.ATTR_STATE) == c.VALUE_ON
-
-    async def group_set_mute(self, group_id: int, state: bool) -> None:
-        """Set the mute state of the group.
-
-        References:
-            4.3.9 Set Group Mute"""
-        await self._connection.command(GroupCommands.group_set_mute(group_id, state))
-
-    async def group_toggle_mute(self, group_id: int) -> None:
-        """Toggle the mute state.
-
-        References:
-            4.3.10 Toggle Group Mute"""
-        await self._connection.command(GroupCommands.group_toggle_mute(group_id))
-
-
-class Heos(SystemCommands, BrowseMixin, GroupMixin, PlayerMixin):
+class Heos(SystemCommands, BrowseMixin, GroupCommands, PlayerMixin):
     """The Heos class provides access to the CLI API."""
 
     @classmethod
